requirements.txt
gradio~=3.12.0
numpy~=1.23.5
torch~=1.13.0
Pillow~=9.3.0
transformers~=4.24.0
opencv-python~=4.6.0.66
